+++ RELEASE_NOTES.md
@@ -7,6 +7,10 @@
 * New transformation: Delete column
 * New transformation: Rename column
 
+### Bugfixes
+
+* Fix broken Pie and Donut charts
+
 ## 0.4.1
 
 Date 2017-01-04